requirements.txt
openai==1.18.0
pandas==2.2.2
python-dotenv==1.0.1
streamlit==1.33.0
